refactor: openAPI schemas default responses
