+++ src/presentation/controllers/articles/CreateArticleController.ts
@@ -1,7 +1,8 @@
 import { CreateArticleUseCase } from 'application/use-cases/articles/CreateArticleUseCase';
-import { Bool, OpenAPIRoute } from 'chanfana';
+import { OpenAPIRoute } from 'chanfana';
 import articleRepository from 'infrastructure/database/repositories/articles';
 import { withErrorHandling } from 'presentation/decorators';
+import { errorResponses, successResponse } from 'presentation/schemas/responses';
 import { z } from 'zod';
 
 export class CreateArticleController extends OpenAPIRoute {
@@ -40,57 +41,37 @@ export class CreateArticleController extends OpenAPIRoute {
                 description: 'Article created successfully',
                 content: {
                     'application/json': {
-                        schema: z.object({
-                            success: Bool(),
-                            result: z.object({
-                                id: z.number(),
-                                title: z.string(),
-                                description: z.string().nullable(),
-                                bodyText: z.string().nullable(),
-                                secondText: z.string().nullable(),
-                                professionalId: z.number().optional(),
-                                author: z.string().optional(),
-                                published: z.string().optional(),
-                                images: z
-                                    .array(
-                                        z.object({
-                                            id: z.number(),
-                                            researchId: z.number(),
-                                            url: z.string().nullable(),
-                                            title: z.string().nullable(),
-                                            description: z.string().nullable()
-                                        })
-                                    )
-                                    .nullable(),
-                                createdAt: z.string(),
-                                updatedAt: z.string()
-                            })
-                        })
+                        schema: z.object(
+                            successResponse(
+                                z.object({
+                                    id: z.number(),
+                                    title: z.string(),
+                                    description: z.string().nullable(),
+                                    bodyText: z.string().nullable(),
+                                    secondText: z.string().nullable(),
+                                    professionalId: z.number().optional(),
+                                    author: z.string().optional(),
+                                    published: z.string().optional(),
+                                    images: z
+                                        .array(
+                                            z.object({
+                                                id: z.number(),
+                                                researchId: z.number(),
+                                                url: z.string().nullable(),
+                                                title: z.string().nullable(),
+                                                description: z.string().nullable()
+                                            })
+                                        )
+                                        .nullable(),
+                                    createdAt: z.string(),
+                                    updatedAt: z.string()
+                                })
+                            )
+                        )
                     }
                 }
             },
-            '400': {
-                description: 'Invalid input',
-                content: {
-                    'application/json': {
-                        schema: z.object({
-                            success: Bool(),
-                            message: z.string()
-                        })
-                    }
-                }
-            },
-            '500': {
-                description: 'Server error',
-                content: {
-                    'application/json': {
-                        schema: z.object({
-                            success: Bool(),
-                            message: z.string()
-                        })
-                    }
-                }
-            }
+            ...errorResponses
         }
     };
