+++ src/presentation/controllers/articles/UpdateArticleController.ts
@@ -1,6 +1,8 @@
 import { UpdateArticleUseCase } from 'application/use-cases/articles/UpdateArticleUseCase';
-import { Bool, OpenAPIRoute } from 'chanfana';
+import { OpenAPIRoute } from 'chanfana';
 import articleRepository from 'infrastructure/database/repositories/articles';
+import { withErrorHandling } from 'presentation/decorators';
+import { errorResponses, successResponse } from 'presentation/schemas/responses';
 import { z } from 'zod';
 
 export class UpdateArticleController extends OpenAPIRoute {
@@ -45,60 +47,41 @@ export class UpdateArticleController extends OpenAPIRoute {
                 description: 'Article updated successfully',
                 content: {
                     'application/json': {
-                        schema: z.object({
-                            success: Bool(),
-                            result: z.object({
-                                id: z.number(),
-                                title: z.string(),
-                                description: z.string().nullable(),
-                                bodyText: z.string().nullable(),
-                                secondText: z.string().nullable(),
-                                createdAt: z.string(),
-                                updatedAt: z.string(),
-                                professionalId: z.number().nullable(),
-                                author: z.string().nullable().optional(),
-                                published: z.string().nullable().optional(),
-                                images: z
-                                    .array(
-                                        z.object({
-                                            id: z.number().nullable(),
-                                            url: z.string().nullable(),
-                                            title: z.string().nullable(),
-                                            description: z.string().nullable(),
-                                            articleId: z.number().nullable()
-                                        })
-                                    )
-                                    .nullable()
-                            })
-                        })
+                        schema: z.object(
+                            successResponse(
+                                z.object({
+                                    id: z.number(),
+                                    title: z.string(),
+                                    description: z.string().nullable(),
+                                    bodyText: z.string().nullable(),
+                                    secondText: z.string().nullable(),
+                                    createdAt: z.string(),
+                                    updatedAt: z.string(),
+                                    professionalId: z.number().nullable(),
+                                    author: z.string().nullable().optional(),
+                                    published: z.string().nullable().optional(),
+                                    images: z
+                                        .array(
+                                            z.object({
+                                                id: z.number().nullable(),
+                                                url: z.string().nullable(),
+                                                title: z.string().nullable(),
+                                                description: z.string().nullable(),
+                                                articleId: z.number().nullable()
+                                            })
+                                        )
+                                        .nullable()
+                                })
+                            )
+                        )
                     }
                 }
             },
-            '400': {
-                description: 'Invalid input',
-                content: {
-                    'application/json': {
-                        schema: z.object({
-                            success: Bool(),
-                            message: z.string()
-                        })
-                    }
-                }
-            },
-            '404': {
-                description: 'Article not found',
-                content: {
-                    'application/json': {
-                        schema: z.object({
-                            success: Bool(),
-                            message: z.string()
-                        })
-                    }
-                }
-            }
+            ...errorResponses
         }
     };
 
+    @withErrorHandling
     async handle(): Promise<object> {
         const data = await this.getValidatedData<typeof this.schema>();
 
@@ -114,58 +97,51 @@ export class UpdateArticleController extends OpenAPIRoute {
             images
         } = data.body;
 
-        try {
-            const updateArticleUseCase = new UpdateArticleUseCase(articleRepository);
+        const updateArticleUseCase = new UpdateArticleUseCase(articleRepository);
 
-            const article = await updateArticleUseCase.execute({
-                id,
-                title,
-                description,
-                bodyText,
-                secondText,
-                professionalId,
-                author,
-                published,
-                images: images
-                    ? images.map((image) => ({
+        const article = await updateArticleUseCase.execute({
+            id,
+            title,
+            description,
+            bodyText,
+            secondText,
+            professionalId,
+            author,
+            published,
+            images: images
+                ? images.map((image) => ({
+                      id: image.id,
+                      url: image.url,
+                      title: image.title,
+                      description: image.description,
+                      articleId: image.articleId
+                  }))
+                : null
+        });
+
+        return {
+            success: true,
+            result: {
+                id: article.id,
+                title: article.title,
+                description: article.description,
+                bodyText: article.bodyText,
+                secondText: article.secondText,
+                author: article.author,
+                professionalId: article.professionalId,
+                published: article.published,
+                images: article.images
+                    ? article.images.map((image) => ({
                           id: image.id,
+                          articleId: image.articleId,
                           url: image.url,
                           title: image.title,
-                          description: image.description,
-                          articleId: image.articleId
+                          description: image.description
                       }))
-                    : null
-            });
-
-            return {
-                success: true,
-                result: {
-                    id: article.id,
-                    title: article.title,
-                    description: article.description,
-                    bodyText: article.bodyText,
-                    secondText: article.secondText,
-                    author: article.author,
-                    professionalId: article.professionalId,
-                    published: article.published,
-                    images: article.images
-                        ? article.images.map((image) => ({
-                              id: image.id,
-                              articleId: image.articleId,
-                              url: image.url,
-                              title: image.title,
-                              description: image.description
-                          }))
-                        : null,
-                    createdAt: article.createdAt.toISOString(),
-                    updatedAt: article.updatedAt.toISOString()
-                }
-            };
-        } catch (error) {
-            return {
-                success: false,
-                message: error.message || 'Article update failed'
-            };
-        }
+                    : null,
+                createdAt: article.createdAt.toISOString(),
+                updatedAt: article.updatedAt.toISOString()
+            }
+        };
     }
 }
