+++ src/presentation/controllers/researchs/DeleteResearchController.ts
@@ -1,6 +1,8 @@
 import { DeleteResearchUseCase } from 'application/use-cases/researchs/DeleteResearchUseCase';
-import { Bool, OpenAPIRoute } from 'chanfana';
+import { OpenAPIRoute } from 'chanfana';
 import researchRepository from 'infrastructure/database/repositories/researchs';
+import { withErrorHandling } from 'presentation/decorators';
+import { errorResponses, simpleSuccessResponse } from 'presentation/schemas/responses';
 import { z } from 'zod';
 
 export class DeleteResearchController extends OpenAPIRoute {
@@ -18,44 +20,26 @@ export class DeleteResearchController extends OpenAPIRoute {
                 description: 'Research deleted successfully',
                 content: {
                     'application/json': {
-                        schema: z.object({
-                            success: Bool()
-                        })
+                        schema: z.object(simpleSuccessResponse)
                     }
                 }
             },
-            '404': {
-                description: 'Research not found',
-                content: {
-                    'application/json': {
-                        schema: z.object({
-                            success: Bool(),
-                            message: z.string()
-                        })
-                    }
-                }
-            }
+            ...errorResponses
         }
     };
 
+    @withErrorHandling
     async handle(): Promise<object> {
         const data = await this.getValidatedData<typeof this.schema>();
 
         const { id } = data.params;
 
-        try {
-            const deleteResearchUseCase = new DeleteResearchUseCase(researchRepository);
+        const deleteResearchUseCase = new DeleteResearchUseCase(researchRepository);
 
-            await deleteResearchUseCase.execute(id);
+        await deleteResearchUseCase.execute(id);
 
-            return {
-                success: true
-            };
-        } catch (error) {
-            return {
-                success: false,
-                message: error.message || 'Research deletion failed'
-            };
-        }
+        return {
+            success: true
+        };
     }
 }
